pyproject.toml
[project]
name = "who-is-my-user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.30.0",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "google-auth>=2.23.0",
    "resend>=0.6.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "argon2-cffi>=23.1.0",
    "requests>=2.31.0",
]
